# docs/openapi-v3.1.json
{
  "components": {
    "schemas": {
      "AccessCodeRequest": {
        "properties": {
          "code": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Code",
            "type": "string"
          }
        },
        "required": [
          "code"
        ],
        "title": "AccessCodeRequest",
        "type": "object"
      },
      "ConfigOption": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "id",
          "label"
        ],
        "title": "ConfigOption",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "JobCreated": {
        "properties": {
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "completed",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          },
          "status_url": {
            "title": "Status Url",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "status",
          "status_url"
        ],
        "title": "JobCreated",
        "type": "object"
      },
      "JobStatus": {
        "properties": {
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "progress": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Progress",
            "type": "integer"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "completed",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "video_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Video Url"
          }
        },
        "required": [
          "job_id",
          "status",
          "progress",
          "message"
        ],
        "title": "JobStatus",
        "type": "object"
      },
      "SessionCreated": {
        "properties": {
          "expires_in_seconds": {
            "title": "Expires In Seconds",
            "type": "integer"
          },
          "token": {
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token",
          "expires_in_seconds"
        ],
        "title": "SessionCreated",
        "type": "object"
      },
      "SessionStatus": {
        "properties": {
          "auth_required": {
            "title": "Auth Required",
            "type": "boolean"
          },
          "authenticated": {
            "title": "Authenticated",
            "type": "boolean"
          }
        },
        "required": [
          "auth_required",
          "authenticated"
        ],
        "title": "SessionStatus",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WebConfig": {
        "properties": {
          "durations": {
            "items": {
              "type": "integer"
            },
            "title": "Durations",
            "type": "array"
          },
          "provider": {
            "enum": [
              "offline",
              "openai"
            ],
            "title": "Provider",
            "type": "string"
          },
          "styles": {
            "items": {
              "$ref": "#/components/schemas/ConfigOption"
            },
            "title": "Styles",
            "type": "array"
          },
          "voices": {
            "items": {
              "$ref": "#/components/schemas/ConfigOption"
            },
            "title": "Voices",
            "type": "array"
          }
        },
        "required": [
          "provider",
          "durations",
          "voices",
          "styles"
        ],
        "title": "WebConfig",
        "type": "object"
      },
      "WebJobRequest": {
        "properties": {
          "duration_seconds": {
            "default": 60,
            "enum": [
              45,
              60,
              90
            ],
            "title": "Duration Seconds",
            "type": "integer"
          },
          "style": {
            "default": "xianxia-cinematic",
            "enum": [
              "xianxia-cinematic",
              "ink-wash",
              "dark-fantasy"
            ],
            "title": "Style",
            "type": "string"
          },
          "topic": {
            "maxLength": 600,
            "minLength": 1,
            "title": "Topic",
            "type": "string"
          },
          "voice": {
            "default": "marin",
            "enum": [
              "marin",
              "onyx"
            ],
            "title": "Voice",
            "type": "string"
          }
        },
        "required": [
          "topic"
        ],
        "title": "WebJobRequest",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Private mobile-first job API for generating narrated vertical videos.",
    "title": "Auto Video Factory Web API",
    "version": "0.3.1"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/config": {
      "get": {
        "operationId": "config_api_config_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebConfig"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Config"
      }
    },
    "/api/jobs": {
      "post": {
        "operationId": "create_job_api_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebJobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobCreated"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Job"
      }
    },
    "/api/jobs/{job_id}": {
      "get": {
        "operationId": "job_status_api_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Job Status"
      }
    },
    "/api/jobs/{job_id}/video": {
      "get": {
        "operationId": "job_video_api_jobs__job_id__video_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Job Video"
      }
    },
    "/api/session": {
      "delete": {
        "operationId": "delete_session_api_session_delete",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Session"
      },
      "get": {
        "operationId": "session_status_api_session_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionStatus"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Session Status"
      },
      "post": {
        "operationId": "create_session_api_session_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccessCodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionCreated"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "description": "Too Many Requests"
          }
        },
        "summary": "Create Session"
      }
    }
  }
}